started cmip6_omz.units, moved from omz_tools
